summary/requirements.txt
requests
pandas
pytest
pylint
pyodbc
python-dotenv
fastparquet
awswrangler
boto3